BINge make chimera handling explicit; allow multiinput
Before, chimeras were being skipped but we were keeping track of them in case we wanted to "fix" them by breaking them apart somehow. That's not needed, so I've removed that tracking to streamline things and make it clear that chimera detection only results in a GMAP path being skipped - that's it.

Also implement the use of FastaCollection to prevent the user from needing to concatenate FASTA files before using the program.