BINge_counter modifed output for DESeq2
Now provides the necessary files for running DESeq2 using Salmon abundance data. Adapts code used by tximport in R to produce similar results. The results aren't identical because tximport seems to do some filtering on the data which is not entirely transparent and I don't think is necessary for BINge.