generate_annotation_table fix 'not really obsoleted'
The condition where a term is absent in the GO.obo but isn't obsolete according to the GO website wasn't being handled correctly. Now, we'll keep the 'not really obsoleted' term, but it will not have ancestor terms found for it. Those can be manually fixed or just accepted as a small (probably inconsequential) sacrifice in the name of having useable results.