evaluate_clustering --beTolerant behaviour
This parameter is necessary to enable the benchmarking process to occur. There are minor differences which can't be rectified without removing some true label genes from the comparison. It's a small enough number to not be of concern, and I believe it's from my gff3_to_fasta not outtputting some sequences lacking CDS' ?