BINge make GMAP alignment cutoff configurable
This represents a pretty large change to the code but is necessary for being properly flexible with cross species clustering. The previous default was always 95% identity cut-off, but we might want to increase that or decrease it in various situations. Benchmarking will help to figure out when and where to do this.