BINge begin overhauling input system
My plan is to make BINge automatically perform the GMAP searches in the way that is expected since it's starting to get complicated. In doing so, I need to create a predictable working environment, and also need to extract the sequences from a GFF3 automatically. This commit includes a lot of the code needed for this.