Revamp processes to hopefully work now
Also cleaned up some unused code i.e., the bin splitter stuff, and updated the testing code.

I _think_ this works now, but it's been a nightmare getting processing to work in a way that 1) exceptions are handled in the main thread, and 2) results are returned to the main thread in order.

There's a chance this implementation encounters deadlocks... but I'm really hoping this actually works now.