BINge force algorithm to use reference seeding
This algorithm design is not really capable of doing things totally de novo. I've hence removed that as an option, and put back in the chimera filter to give a rough estimate on how well this algorithm can perform.

I'm going to develop a different approach on another git branch and see how that works. If it's better, it'll replace this main branch if I can manage that without ruining the repo.