Detect changes to input arguments and error out
A current issue with the way BINge using a working directory is that it assumes no one has messed with things. That flexibility is good for me to use since I know how to hack it without ruining things, but I might make mistakes and others might not do it right. These changes should hopefully make it very difficult for someone to re-run an analysis in a way that makes results unreliable.