Use genome models for cluster; evaluation commit
Discovered an issue that occurs in the vertebrates analysis. If sequences identical to the original ones used for pre-seeding exist but have different IDs to the GFF3, then we can end up with clustering of these occurring separately. This issue means I must re-run any evaluations that relied upon pre-seeding with this commit.