Full removal of Various_scripts requirements
All code has been copy pasted into appropriate existing modules or new modules were created to house the code. I didn't want to do this originally for code stability, but there's a lot of annoying extra dependencies that come with importing Various_scripts and I want BINge to be as easy to install as possible.