Cessation of branch
Many changes have been made, but these have been a waste of time. This solution is not practical since I'd need to accommodate BLAST, salmon, and other things to handle duplicated IDs. A more fundamental fix is needed since I'm having to adjust the code base radically in ways that it is fragile to.